fix: missing expense deletion validation for existance, return 404

+++ src/modules/expenses/delete-expense.ts
@@ -6,15 +6,23 @@ import { resourceOwnerId } from "../../middlewares/auth";
 const deleteExpense = async (c: Context) => {
     const id = c.req.param("id");
 
+    const expenseExist = await prisma.expense.findUnique({
+        where: {
+            id,
+            user: {
+                id: {
+                    equals: resourceOwnerId(c)
+                }
+            }
+        }
+    })
+
+    if (!expenseExist) return c.json({ error: "Expense not found" }, 404);
+
+
     try {
         await prisma.expense.delete({
-            where: {
-                id, user: {
-                    id: {
-                        equals: resourceOwnerId(c)
-                    }
-                }
-            },
+            where: { id },
         });
 
         return c.json({ message: "Expense deleted successfully" });
